refactor: extract client request resolution
